+++ app/src/main/java/com/infomaniak/mail/ui/ThemedActivity.kt
@@ -15,18 +15,22 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main
+package com.infomaniak.mail.ui
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
+import androidx.appcompat.app.AppCompatActivity
 import com.infomaniak.mail.R
 
-class NewFolderDialog : DialogFragment() {
+open class ThemedActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        val isPink = false
 
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        return inflater.inflate(R.layout.dialog_new_folder, container, false)
+        val theme = when {
+            isPink -> R.style.AppTheme_Pink
+            else -> R.style.AppTheme_Blue
+        }
+        setTheme(theme)
+
+        super.onCreate(savedInstanceState)
     }
 }
