+++ app/src/main/java/com/infomaniak/mail/ui/main/contact/ContactFragment.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main
+package com.infomaniak.mail.ui.main.contact
 
 import android.os.Bundle
 import android.view.LayoutInflater
