+++ app/src/main/java/com/infomaniak/mail/ui/main/menu/ManageMailAddressFragment.kt
@@ -27,6 +27,6 @@ import com.infomaniak.mail.R
 class ManageMailAddressFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        return inflater.inflate(R.layout.fragment_manager_mail_address, container, false)
+        return inflater.inflate(R.layout.fragment_manage_mail_address, container, false)
     }
 }
