+++ app/src/main/java/com/infomaniak/mail/ui/main/newMessage/NewMessageActivity.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main.newmessage
+package com.infomaniak.mail.ui.main.newMessage
 
 import android.os.Bundle
 import androidx.activity.viewModels
@@ -31,9 +31,9 @@ import com.infomaniak.mail.data.models.MessagePriority
 import com.infomaniak.mail.data.models.MessagePriority.getPriority
 import com.infomaniak.mail.data.models.Recipient
 import com.infomaniak.mail.databinding.ActivityNewMessageBinding
-import com.infomaniak.mail.ui.main.MainViewModel
-import com.infomaniak.mail.ui.main.ThemedActivity
-import com.infomaniak.mail.ui.main.newmessage.NewMessageActivity.EditorAction.*
+import com.infomaniak.mail.ui.MainViewModel
+import com.infomaniak.mail.ui.ThemedActivity
+import com.infomaniak.mail.ui.main.newMessage.NewMessageActivity.EditorAction.*
 import io.realm.kotlin.ext.realmListOf
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
