+++ app/src/main/java/com/infomaniak/mail/ui/main/search/SearchFragment.kt
@@ -15,35 +15,18 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main.thread
+package com.infomaniak.mail.ui.main.search
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import androidx.fragment.app.Fragment
 import com.infomaniak.mail.R
-import com.infomaniak.mail.databinding.BottomSheetReplyBinding
-import com.infomaniak.mail.utils.notYetImplemented
 
-open class ReplyBottomSheetDialog : BottomSheetDialogFragment() {
-
-    lateinit var binding: BottomSheetReplyBinding
+class SearchFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        return BottomSheetReplyBinding.inflate(inflater, container, false).also { binding = it }.root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) = with(binding) {
-        super.onViewCreated(view, savedInstanceState)
-
-        mainActions.setOnItemClickListener { id: Int ->
-            when (id) {
-                R.id.actionReply -> notYetImplemented()
-                R.id.actionReplyAll -> notYetImplemented()
-            }
-            findNavController().popBackStack()
-        }
+        return inflater.inflate(R.layout.fragment_search, container, false)
     }
 }
