+++ app/src/main/java/com/infomaniak/mail/ui/main/thread/actions/ActionItemView.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main.thread
+package com.infomaniak.mail.ui.main.thread.actions
 
 import android.content.Context
 import android.content.res.ColorStateList
