+++ app/src/main/java/com/infomaniak/mail/ui/main/user/SwitchUserFragment.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.main.menu.user
+package com.infomaniak.mail.ui.main.user
 
 import android.content.Intent
 import android.os.Bundle
@@ -33,8 +33,8 @@ import com.infomaniak.mail.data.cache.mailboxInfos.MailboxController
 import com.infomaniak.mail.data.models.Mailbox
 import com.infomaniak.mail.databinding.FragmentSwitchUserBinding
 import com.infomaniak.mail.ui.LoginActivity
-import com.infomaniak.mail.ui.main.MainViewModel
-import com.infomaniak.mail.ui.main.menu.user.SwitchUserAccountsAdapter.UiAccount
+import com.infomaniak.mail.ui.MainViewModel
+import com.infomaniak.mail.ui.main.user.SwitchUserAccountsAdapter.UiAccount
 import com.infomaniak.mail.utils.AccountUtils
 import com.infomaniak.mail.utils.sortMailboxes
 import io.realm.kotlin.notifications.ResultsChange
