+++ app/src/main/res/layout/activity_new_message.xml
@@ -21,7 +21,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:clipChildren="false"
-    tools:context=".ui.main.newmessage.NewMessageActivity">
+    tools:context=".ui.main.newMessage.NewMessageActivity">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
