+++ app/src/main/res/layout/bottom_sheet_detailed_contact.xml
@@ -27,7 +27,7 @@
         android:layout_height="wrap_content"
         android:orientation="vertical"
         android:paddingBottom="@dimen/marginStandardMedium"
-        tools:context=".ui.main.thread.MessageActionsBottomSheetDialog">
+        tools:context=".ui.main.thread.actions.MessageActionsBottomSheetDialog">
 
         <include layout="@layout/view_bottom_sheet_separator" />
 
@@ -86,7 +86,7 @@
 
         </androidx.constraintlayout.widget.ConstraintLayout>
 
-        <com.infomaniak.mail.ui.main.thread.ActionItemView
+        <com.infomaniak.mail.ui.main.thread.actions.ActionItemView
             android:id="@+id/writeMail"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
@@ -95,14 +95,14 @@
             app:text="@string/contactActionWriteEmail"
             app:visibleDivider="false" />
 
-        <com.infomaniak.mail.ui.main.thread.ActionItemView
+        <com.infomaniak.mail.ui.main.thread.actions.ActionItemView
             android:id="@+id/addToContacts"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             app:icon="@drawable/ic_contact_action_add_user"
             app:text="@string/contactActionAddToContacts" />
 
-        <com.infomaniak.mail.ui.main.thread.ActionItemView
+        <com.infomaniak.mail.ui.main.thread.actions.ActionItemView
             android:id="@+id/copyAddress"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
