+++ app/src/main/res/layout/dialog_new_folder.xml
@@ -20,7 +20,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.NewFolderDialog">
+    tools:context=".ui.main.menu.NewFolderDialog">
 
     <TextView
         android:layout_width="wrap_content"
