+++ app/src/main/res/layout/fragment_bottom_sheet_message_actions.xml
@@ -20,7 +20,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.thread.MessageActionsBottomSheetDialog">
+    tools:context=".ui.main.thread.actions.MessageActionsBottomSheetDialog">
 
     <TextView
         android:layout_width="wrap_content"
