+++ app/src/main/res/layout/fragment_manage_mail_address.xml
