+++ app/src/main/res/layout/fragment_new_message.xml
@@ -20,7 +20,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".ui.main.newmessage.NewMessageFragment">
+    tools:context=".ui.main.newMessage.NewMessageFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintLayout"
