+++ app/src/main/res/layout/fragment_select_settings.xml
@@ -20,7 +20,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.main.ContactFragment">
+    tools:context=".ui.main.contact.ContactFragment">
 
     <TextView
         android:layout_width="wrap_content"
