+++ app/src/main/res/navigation/new_message_navigation.xml
@@ -23,7 +23,7 @@
 
     <fragment
         android:id="@+id/newMessageFragment"
-        android:name="com.infomaniak.mail.ui.main.newmessage.NewMessageFragment"
+        android:name="com.infomaniak.mail.ui.main.newMessage.NewMessageFragment"
         android:label="fragment_new_message"
         tools:layout="@layout/fragment_new_message" />
 </navigation>
